Inte-platform/platform-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-multipart==0.0.6
pydantic[email]==2.5.2
pydantic-settings==2.1.0
pyyaml==6.0.1
httpx==0.25.2
prometheus-client==0.19.0
